+++ subprojects/execution/src/test/groovy/org/gradle/internal/execution/fingerprint/impl/DefaultInputFingerprinterTest.groovy
@@ -20,10 +20,11 @@ import com.google.common.collect.ImmutableSortedMap
 import org.gradle.api.file.FileCollection
 import org.gradle.api.tasks.FileNormalizer
 import org.gradle.internal.execution.UnitOfWork
-import org.gradle.internal.execution.UnitOfWork.FileValueSupplier
 import org.gradle.internal.execution.fingerprint.FileCollectionFingerprinter
 import org.gradle.internal.execution.fingerprint.FileCollectionFingerprinterRegistry
 import org.gradle.internal.execution.fingerprint.FileNormalizationSpec
+import org.gradle.internal.execution.fingerprint.InputFingerprinter.InputVisitor
+import org.gradle.internal.execution.fingerprint.InputFingerprinter.FileValueSupplier
 import org.gradle.internal.execution.fingerprint.InputFingerprinter.Result
 import org.gradle.internal.fingerprint.CurrentFileCollectionFingerprint
 import org.gradle.internal.fingerprint.DirectorySensitivity
@@ -33,7 +34,7 @@ import spock.lang.Specification
 
 import java.util.function.Consumer
 
-import static org.gradle.internal.execution.UnitOfWork.InputPropertyType.NON_INCREMENTAL
+import static org.gradle.internal.execution.fingerprint.InputFingerprinter.InputPropertyType.NON_INCREMENTAL
 
 class DefaultInputFingerprinterTest extends Specification {
     def work = Mock(UnitOfWork)
@@ -112,7 +113,7 @@ class DefaultInputFingerprinterTest extends Specification {
         ImmutableSortedMap<String, ValueSnapshot> previousValueSnapshots = ImmutableSortedMap.of(),
         ImmutableSortedMap<String, ValueSnapshot> knownValueSnapshots = ImmutableSortedMap.of(),
         ImmutableSortedMap<String, CurrentFileCollectionFingerprint> knownFingerprints = ImmutableSortedMap.of(),
-        Consumer<UnitOfWork.InputVisitor> inputs
+        Consumer<InputVisitor> inputs
     ) {
         inputFingerprinter.fingerprintInputProperties(inputs, previousValueSnapshots, knownValueSnapshots, knownFingerprints)
     }
